requirements.txt
gmpy2
google-re2
gravis
jsonplus
matplotlib # Graph plotting
networkx
numpy
parse
python-dateutil
requests
scipy # Graph plotting
shapely # Shape math
sympy
z3-solver